Started Afresh
I faced many challenges with my local repo pushing it online and that was
squarely caused by me. I forked the Memory Project Repo and cloned it to
my local machine. I started off following Matt Cranford's guide and was
carried away with anything called git because the project is too beyond
my scope of understanding into FEND journey so far. Even while I was
on the guide, I almost gave up. I had to share it on Slack before I got
encouraging messages from others to fight on. Truly, it's a torn and my
desire is to get to the end of this journey. I will suggest that Udacity
review the curriculum and make it easier to build as you watch the
lectures rather than waiting a long period to start a project. I am yet
to submit my project at least for the first review but I must confess
it's never a simple package. I have deleted the forked repo and tried
about to push my local files to the new repo. I hope it works this time.
Thanks to "drunkenkismet FEND" for his advise on this I am about doing.